+++ appGUI/MainGUI.py
@@ -2893,7 +2893,7 @@ class MainGUI(QtWidgets.QMainWindow):
 
                 # Properties Tool
                 if key == QtCore.Qt.Key_P:
-                    self.app.properties_tool.run()
+                    self.app.properties_tool.run(toggle=True)
                     return
 
                 # Change Units